Move and rename circuit_constructor dirs and files
