+++ cpp/src/barretenberg/benchmark/honk_bench/standard_plonk.bench.cpp
@@ -1,5 +1,5 @@
 #include "barretenberg/benchmark/honk_bench/benchmark_utilities.hpp"
-#include "barretenberg/proof_system/circuit_constructors/standard_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/standard_circuit_builder.hpp"
 #include "barretenberg/plonk/composer/standard_composer.hpp"
 
 using namespace benchmark;
