+++ cpp/src/barretenberg/benchmark/honk_bench/ultra_honk.bench.cpp
@@ -1,6 +1,6 @@
 #include <benchmark/benchmark.h>
 
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/honk/composer/ultra_composer.hpp"
 #include "barretenberg/benchmark/honk_bench/benchmark_utilities.hpp"
