+++ cpp/src/barretenberg/benchmark/honk_bench/ultra_plonk.bench.cpp
@@ -1,5 +1,5 @@
 #include "barretenberg/benchmark/honk_bench/benchmark_utilities.hpp"
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/plonk/composer/ultra_composer.hpp"
 
 using namespace benchmark;
