+++ cpp/src/barretenberg/honk/composer/standard_composer.test.cpp
@@ -3,7 +3,7 @@
 #include <vector>
 #include <gtest/gtest.h>
 
-#include "barretenberg/proof_system/circuit_constructors/standard_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/standard_circuit_builder.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
 #include "barretenberg/honk/composer/standard_composer.hpp"
