+++ cpp/src/barretenberg/plonk/composer/standard_composer.test.cpp
@@ -3,7 +3,7 @@
 #include "barretenberg/crypto/generators/generator_data.hpp"
 #include "barretenberg/plonk/composer/standard_composer.hpp"
 #include "barretenberg/plonk/proof_system/proving_key/serialize.hpp"
-#include "barretenberg/proof_system/circuit_constructors/standard_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/standard_circuit_builder.hpp"
 
 using namespace barretenberg;
 using namespace proof_system;
