+++ cpp/src/barretenberg/plonk/composer/turbo_composer.test.cpp
@@ -4,7 +4,7 @@
 #include "barretenberg/crypto/generators/fixed_base_scalar_mul.hpp"
 #include "barretenberg/plonk/composer/turbo_composer.hpp"
 #include "barretenberg/plonk/proof_system/proving_key/serialize.hpp"
-#include "barretenberg/proof_system/circuit_constructors/turbo_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/turbo_circuit_builder.hpp"
 
 using namespace barretenberg;
 using namespace proof_system;
