+++ cpp/src/barretenberg/proof_system/circuit_builder/circuit_builder_base.hpp
