+++ cpp/src/barretenberg/proof_system/circuit_builder/standard_circuit_builder.cpp
@@ -1,4 +1,4 @@
-#include "standard_circuit_constructor.hpp"
+#include "standard_circuit_builder.hpp"
 #include <unordered_set>
 #include <unordered_map>
