+++ cpp/src/barretenberg/proof_system/circuit_builder/turbo_circuit_builder.cpp
@@ -1,4 +1,4 @@
-#include "turbo_circuit_constructor.hpp"
+#include "turbo_circuit_builder.hpp"
 #include "barretenberg/common/slab_allocator.hpp"
 #include "barretenberg/numeric/bitop/get_msb.hpp"
