+++ cpp/src/barretenberg/proof_system/circuit_builder/turbo_circuit_builder.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <array>
-#include "circuit_constructor_base.hpp"
+#include "circuit_builder_base.hpp"
 #include "barretenberg/proof_system/types/circuit_type.hpp"
 #include "barretenberg/proof_system/types/merkle_hash_type.hpp"
 #include "barretenberg/proof_system/types/pedersen_commitment_type.hpp"
