+++ cpp/src/barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "barretenberg/proof_system/arithmetization/arithmetization.hpp"
 #include "barretenberg/plonk/proof_system/types/polynomial_manifest.hpp"
-#include "circuit_constructor_base.hpp"
+#include "circuit_builder_base.hpp"
 #include "barretenberg/plonk/proof_system/constants.hpp"
 #include "barretenberg/proof_system/types/merkle_hash_type.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
