+++ cpp/src/barretenberg/proof_system/circuit_builder/ultra_circuit_builder.test.cpp
@@ -1,4 +1,4 @@
-#include "ultra_circuit_constructor.hpp"
+#include "ultra_circuit_builder.hpp"
 #include "barretenberg/crypto/generators/generator_data.hpp"
 #include <gtest/gtest.h>
