+++ cpp/src/barretenberg/stdlib/encryption/schnorr/schnorr.test.cpp
@@ -3,7 +3,7 @@
 #include "schnorr.hpp"
 #include "barretenberg/crypto/pedersen_commitment/pedersen.hpp"
 #include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/stdlib/primitives/point/point.hpp"
 
 namespace proof_system::test_stdlib_schnorr {
