+++ cpp/src/barretenberg/stdlib/hash/blake2s/blake2s.test.cpp
@@ -2,7 +2,7 @@
 #include "blake2s_plookup.hpp"
 #include <gtest/gtest.h>
 #include "barretenberg/crypto/blake2s/blake2s.hpp"
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 
 using namespace barretenberg;
 using namespace proof_system::plonk::stdlib;
