+++ cpp/src/barretenberg/stdlib/merkle_tree/hash.test.cpp
@@ -2,7 +2,7 @@
 #include "memory_tree.hpp"
 #include <gtest/gtest.h>
 
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/stdlib/merkle_tree/membership.hpp"
 
 namespace proof_system::stdlib_merkle_tree_hash_test {
