+++ cpp/src/barretenberg/stdlib/primitives/memory/rom_table.test.cpp
@@ -3,7 +3,7 @@
 
 #include "rom_table.hpp"
 #include "barretenberg/numeric/random/engine.hpp"
-#include "barretenberg/proof_system/circuit_constructors/ultra_circuit_constructor.hpp"
+#include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 
 namespace test_stdlib_rom_array {
 using namespace barretenberg;
